+++ R/onLoad.R
@@ -1,14 +1,7 @@
-#' @title Loads the iDAI.field Categories
-#' @description This function loads the iDAI.field categories with the default values.
-#' @param libname The name of the library.
-#' @param pkgname The name of the package.
-#' @return A list of options.
-#' @export
-#' @examples
-#' \dontrun{
-#' idaifield_categories <- .onLoad("mylib", "idaifieldR")
-#' }
-.onLoad <- function(libname, pkgname) {
+#' @title onLoad / onAttach
+#' @param lib The name of the library.
+#' @param pkg The name of the package.
+.onLoad <- function(lib, pkg) {
   # The options are set in a list.
   options(
     # The list contains a list of options.
@@ -27,4 +20,8 @@
     )
     )
   )
+}
+
+.onAttach <- function(lib, pkg) {
+  check_idaifieldr_version(getNamespaceVersion('idaifieldR'))
 }