+++ tests/testthat/test-simplify_idaifield_resources.R
@@ -111,6 +111,11 @@ test_that("runs without uidlist", {
 
 suppressMessages(simple <- simplify_idaifield(test_docs))
 
+
+test_that("attaches attribute for duplicate names", {
+  expect_contains(names(attributes(simple)), "duplicate_names")
+})
+
 test_that("contains no special config names in columns", {
   expect_false(any(grepl(":", names(simple))))
 })